Send delivery receipt immediately after receiving send response
Previously it would wait for the message ID to come from the echo, but
the echo isn't reliable, so need to do this instead.

Also added a TODO to store the offline threading IDs in the database to
avoid some of the problems that not receiving the message ID can cause.