thread fetch: search backwards to find next good timestamp
If we get a response error, then try to fetch the threads in the page
one-at-a-time.

Once we arrive at a failing thread, find a timestamp earlier in history
that works.

Note that this may cause some threads to be missed if they come close
behind the failing thread, but since failures of this sort are probably
rare, that's an acceptable tradeoff.

Signed-off-by: Sumner Evans <sumner@beeper.com>