Refactor MQTT error handling (#306)
* Offload MQTT reconnection handling to client

This removes a bunch of logic we have added to handle MQTT connection
issues and instead utilises the clients builtin mechanisms. Achieved
by overwriting the keepalive value temporarily while doing send/receive
to force quicker detection of connection issues.

Needs testing, but looks like it could work!

* Simplify reconnecting logic and make retry iterations quicker

* Increase MQTT timeout to 3 minutes

This is high, but gives plenty of buffer for proxy issues to be fixed
or rotated. Idea here is to get things working consistently and we can
later reduce this value for a more aggressive target.

* Upgrade mautrix to v0.19.15

* Reset MQTT error retry counter after 1h stable